changed int port to unsigned short port

+++ inc/Socket.hpp
@@ -23,9 +23,9 @@ class Socket
 	void			   set_socket_non_blocking();
 
   public:
-	Socket(int domain, int port, int type, int protocol, u_long interface);
+	Socket(int domain, unsigned short port, int type, int protocol, u_long interface);
 	int				   get_sock_id() const;
-	unsigned short int get_port() const;
+	unsigned short get_port() const;
 
 	class SocketException : public std::exception
 	{
